# pkg/models/client.go
package models

//
//type Client struct {
//	gorm.Model
//	Name  string       `gorm:"size:255"`
//	Quota int          // Tarrificator quota
//	Users []ClientUser // One-to-many relationship with ClientUser
//}